% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkgpa.R
\name{bulkgpa}
\alias{bulkgpa}
\title{Bulk calculate gpas. Some pre-cleaning in excel may be required.}
\usage{
bulkgpa(
  a = transfer_courses_file,
  b = transfer_hours_file,
  d = UH_courses,
  e = content_areas
)
}
\arguments{
\item{a}{transfer course file from peoplesoft}

\item{b}{fransfer hours file from peoplesoft}

\item{d}{UH courses file from peoplesoft}

\item{e}{Content subjects}
}
\value{
summary of gpa and summary of hours taken
}
\description{
Bulk calculate gpas. Some pre-cleaning in excel may be required.
}
\examples{
bulkgpa(a,b,d)
}
