% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difference.R
\name{difference}
\alias{difference}
\title{Examine the difference between newer datasets}
\usage{
difference(a, b)
}
\arguments{
\item{a}{smaller datafile}

\item{b}{larger datafile}
}
\value{
the row differences that remain
}
\description{
Examine the difference between newer datasets
}
