% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeme.R
\name{mergeme}
\alias{mergeme}
\title{merge two files}
\usage{
mergeme(b, d, e)
}
\arguments{
\item{b}{larger dataset, if there is a difference}

\item{d}{smaller dataset}

\item{e}{id field for original datasets in common. Should be surrounded by quotes.}
}
\value{
a merged set with inner, outer, and all joins.
}
\description{
merge two files
}
