% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/professionals.R
\name{professionals}
\alias{professionals}
\title{Identify professional certs}
\usage{
professionals(a = filename)
}
\arguments{
\item{a}{= filename}
}
\value{
file with only professional certifications
}
\description{
Identify professional certs
}
