% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_column.R
\name{split_column}
\alias{split_column}
\title{Split a column into two columns}
\usage{
split_column(v, w, x, y, z)
}
\arguments{
\item{v}{data frame}

\item{w}{column number}

\item{x}{regex in quotes}

\item{y}{name of first split column}

\item{z}{name of second split column}
}
\value{
dataframe with split columns and original data frame
}
\description{
Split a column into two columns
}
