% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unequal_dataframe.R
\name{unequal_dataframe}
\alias{unequal_dataframe}
\title{An unequal dataframe with NAs to make up the difference}
\usage{
unequal_dataframe(x, y)
}
\arguments{
\item{x}{smaller}

\item{y}{larger datafile}
}
\value{
a dataframe of unequal lengths with NAs
}
\description{
An unequal dataframe with NAs to make up the difference
}
