% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difference.R
\name{difference}
\alias{difference}
\title{examine the difference between newer datasets}
\usage{
difference(
  a = smaller_datafile,
  b = larger_datafile,
  d = column_number_for_id_1,
  p = column_for_id_2,
  m = column_for_index1,
  n = column_number_for_index_2
)
}
\arguments{
\item{a}{smaller data frame}

\item{b}{larger data frame}

\item{d}{column number for first data frame ID}

\item{p}{column number for second data frame ID}

\item{m}{column number for first index number}

\item{n}{column number for second index number}
}
\value{
a data set of rows that remain after comparison
}
\description{
examine the difference between newer datasets
}
